Fixed some compilation warnings and added a few comments.

+++ dsd_filters.c
@@ -112,5 +112,5 @@ dsd_input_filter(short sample, int mode)
   for (i = 0; i <= zeros; i++)
     sum += (coeffs[i] * v[i]);
 
-  return sum / ngain; // filtered sample out
+  return (short)(sum / ngain); // filtered sample out
 }
